+++ lib/td_dd_web/router.ex
@@ -129,7 +129,8 @@ defmodule TdDdWeb.Router do
     end
 
     resources("/grant_requests", GrantRequestController, only: [:index, :show, :delete]) do
-      resources("/approvals", ApprovalController, only: [:create])
+      resources("/approvals", GrantRequestApprovalController, only: [:create], name: "approval")
+      resources("/status", GrantRequestStatusController, only: [:create], name: "status")
     end
   end