+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-4076] Support for grant request approval
 - [TD-4113] Jobs change status is included in Audit Events
+- [TD-4077] Grant request processing workflow support
 
 ### Changed